+++ scala-script-cli/src/scala-script-module.ts
@@ -7,6 +7,8 @@ import {
   LangiumDocument,
   AstNode,
   URI,
+  IndentationAwareTokenBuilder,
+  IndentationAwareLexer,
 } from 'langium'
 import {
   createDefaultModule,
@@ -55,6 +57,20 @@ export const ScalaScriptSharedModule: Module<ScalaScriptSharedServices, DeepPart
  * selected services, while the custom services must be fully specified.
  */
 export const ScalaScriptModule: Module<ScalaScriptServices, PartialLangiumServices & ScalaScriptAddedServices> = {
+  parser: {
+    TokenBuilder: () =>
+      new IndentationAwareTokenBuilder({
+        indentTokenName: 'INDENT',
+        dedentTokenName: 'DEDENT',
+        whitespaceTokenName: 'WS',
+        ignoreIndentationDelimiters: [
+          ['[', ']'],
+          ['{', '}'],
+          ['(', ')'],
+        ],
+      }),
+    Lexer: services => new IndentationAwareLexer(services),
+  },
   validation: {
     ScalaScriptValidator: () => new ScalaScriptValidator(),
   },